requirements.txt
gunicorn==19.6.0
daphne==1.3.0
pip==9.0.1
Django==1.11.2
djangorestframework==3.6.3
django-cors-headers==1.1.0
requests==2.18.1
psycopg2==2.7.1
django-redis==4.8.0
channels==1.1.6
asgi-redis==1.4.2
daphne==1.3.0
simplejson==3.11.1
